# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Financial Fraud Detection System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type TransactionRiskLevel =
  (typeof TransactionRiskLevel)[keyof typeof TransactionRiskLevel];

export const TransactionRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  pending: "pending",
  approved: "approved",
  flagged: "flagged",
  blocked: "blocked",
} as const;

export interface Transaction {
  id: number;
  transactionId: string;
  userId: string;
  userName: string;
  amount: number;
  currency: string;
  merchantName: string;
  merchantCategory: string;
  location: string;
  ipAddress?: string;
  deviceType?: string;
  /** Risk score 0-100 */
  riskScore: number;
  riskLevel: TransactionRiskLevel;
  status: TransactionStatus;
  flagReasons: string[];
  createdAt: string;
}

export interface TransactionListResponse {
  transactions: Transaction[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type AlertSeverity = (typeof AlertSeverity)[keyof typeof AlertSeverity];

export const AlertSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type AlertStatus = (typeof AlertStatus)[keyof typeof AlertStatus];

export const AlertStatus = {
  open: "open",
  reviewed: "reviewed",
  dismissed: "dismissed",
  escalated: "escalated",
} as const;

export interface Alert {
  id: number;
  alertType: string;
  severity: AlertSeverity;
  status: AlertStatus;
  transactionId: number;
  transaction?: Transaction;
  description: string;
  reviewNote?: string;
  reviewedBy?: string;
  reviewedAt?: string;
  createdAt: string;
}

export interface AlertListResponse {
  alerts: Alert[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type ReviewAlertRequestStatus =
  (typeof ReviewAlertRequestStatus)[keyof typeof ReviewAlertRequestStatus];

export const ReviewAlertRequestStatus = {
  reviewed: "reviewed",
  dismissed: "dismissed",
  escalated: "escalated",
} as const;

export interface ReviewAlertRequest {
  status: ReviewAlertRequestStatus;
  reviewNote?: string;
  reviewedBy?: string;
}

export type AnalyticsSummaryRiskDistribution = {
  low: number;
  medium: number;
  high: number;
  critical: number;
};

export interface AnalyticsSummary {
  totalTransactions: number;
  flaggedTransactions: number;
  blockedTransactions: number;
  totalAlertsOpen: number;
  fraudRate: number;
  totalAmountFlagged: number;
  riskDistribution: AnalyticsSummaryRiskDistribution;
  recentAlertCount: number;
}

export interface TrendDataPoint {
  date: string;
  totalTransactions: number;
  flaggedCount: number;
  blockedCount: number;
  fraudRate: number;
}

export type ListTransactionsParams = {
  page?: number;
  limit?: number;
  riskLevel?: ListTransactionsRiskLevel;
  status?: ListTransactionsStatus;
};

export type ListTransactionsRiskLevel =
  (typeof ListTransactionsRiskLevel)[keyof typeof ListTransactionsRiskLevel];

export const ListTransactionsRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type ListTransactionsStatus =
  (typeof ListTransactionsStatus)[keyof typeof ListTransactionsStatus];

export const ListTransactionsStatus = {
  pending: "pending",
  approved: "approved",
  flagged: "flagged",
  blocked: "blocked",
} as const;

export type ListAlertsParams = {
  status?: ListAlertsStatus;
  page?: number;
  limit?: number;
};

export type ListAlertsStatus =
  (typeof ListAlertsStatus)[keyof typeof ListAlertsStatus];

export const ListAlertsStatus = {
  open: "open",
  reviewed: "reviewed",
  dismissed: "dismissed",
  escalated: "escalated",
} as const;

export type GetFraudTrendsParams = {
  days?: number;
};