# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Financial Fraud Detection System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: transactions
    description: Transaction operations
  - name: alerts
    description: Fraud alert operations
  - name: analytics
    description: Analytics and statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions
      description: Returns paginated list of transactions with fraud risk scores
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: riskLevel
          schema:
            type: string
            enum: [low, medium, high, critical]
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, flagged, blocked]
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"

  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get transaction by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List fraud alerts
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [open, reviewed, dismissed, escalated]
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of fraud alerts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertListResponse"

  /alerts/{id}/review:
    patch:
      operationId: reviewAlert
      tags: [alerts]
      summary: Review a fraud alert
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewAlertRequest"
      responses:
        "200":
          description: Alert updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /analytics/summary:
    get:
      operationId: getAnalyticsSummary
      tags: [analytics]
      summary: Get analytics summary
      responses:
        "200":
          description: Analytics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"

  /analytics/trends:
    get:
      operationId: getFraudTrends
      tags: [analytics]
      summary: Get fraud trends over time
      parameters:
        - in: query
          name: days
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Fraud trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendDataPoint"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Transaction:
      type: object
      properties:
        id:
          type: integer
        transactionId:
          type: string
        userId:
          type: string
        userName:
          type: string
        amount:
          type: number
        currency:
          type: string
        merchantName:
          type: string
        merchantCategory:
          type: string
        location:
          type: string
        ipAddress:
          type: string
        deviceType:
          type: string
        riskScore:
          type: number
          description: Risk score 0-100
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [pending, approved, flagged, blocked]
        flagReasons:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - transactionId
        - userId
        - userName
        - amount
        - currency
        - merchantName
        - merchantCategory
        - location
        - riskScore
        - riskLevel
        - status
        - flagReasons
        - createdAt

    TransactionListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - transactions
        - total
        - page
        - limit
        - totalPages

    Alert:
      type: object
      properties:
        id:
          type: integer
        alertType:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [open, reviewed, dismissed, escalated]
        transactionId:
          type: integer
        transaction:
          $ref: "#/components/schemas/Transaction"
        description:
          type: string
        reviewNote:
          type: string
        reviewedBy:
          type: string
        reviewedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - alertType
        - severity
        - status
        - transactionId
        - description
        - createdAt

    AlertListResponse:
      type: object
      properties:
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/Alert"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - alerts
        - total
        - page
        - limit
        - totalPages

    ReviewAlertRequest:
      type: object
      properties:
        status:
          type: string
          enum: [reviewed, dismissed, escalated]
        reviewNote:
          type: string
        reviewedBy:
          type: string
      required:
        - status

    AnalyticsSummary:
      type: object
      properties:
        totalTransactions:
          type: integer
        flaggedTransactions:
          type: integer
        blockedTransactions:
          type: integer
        totalAlertsOpen:
          type: integer
        fraudRate:
          type: number
        totalAmountFlagged:
          type: number
        riskDistribution:
          type: object
          properties:
            low:
              type: integer
            medium:
              type: integer
            high:
              type: integer
            critical:
              type: integer
          required:
            - low
            - medium
            - high
            - critical
        recentAlertCount:
          type: integer
      required:
        - totalTransactions
        - flaggedTransactions
        - blockedTransactions
        - totalAlertsOpen
        - fraudRate
        - totalAmountFlagged
        - riskDistribution
        - recentAlertCount

    TrendDataPoint:
      type: object
      properties:
        date:
          type: string
        totalTransactions:
          type: integer
        flaggedCount:
          type: integer
        blockedCount:
          type: integer
        fraudRate:
          type: number
      required:
        - date
        - totalTransactions
        - flaggedCount
        - blockedCount
        - fraudRate