# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Financial Fraud Detection System API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSeverity } from "./alertSeverity";
import type { AlertStatus } from "./alertStatus";
import type { Transaction } from "./transaction";

export interface Alert {
  id: number;
  alertType: string;
  severity: AlertSeverity;
  status: AlertStatus;
  transactionId: number;
  transaction?: Transaction;
  description: string;
  reviewNote?: string;
  reviewedBy?: string;
  reviewedAt?: Date;
  createdAt: Date;
}