# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Financial Fraud Detection System API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyticsSummaryRiskDistribution } from "./analyticsSummaryRiskDistribution";

export interface AnalyticsSummary {
  totalTransactions: number;
  flaggedTransactions: number;
  blockedTransactions: number;
  totalAlertsOpen: number;
  fraudRate: number;
  totalAmountFlagged: number;
  riskDistribution: AnalyticsSummaryRiskDistribution;
  recentAlertCount: number;
}