# lib/api-zod/src/generated/types/reviewAlertRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Financial Fraud Detection System API
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewAlertRequestStatus } from "./reviewAlertRequestStatus";

export interface ReviewAlertRequest {
  status: ReviewAlertRequestStatus;
  reviewNote?: string;
  reviewedBy?: string;
}