# lib/api-zod/src/generated/types/trendDataPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Financial Fraud Detection System API
 * OpenAPI spec version: 0.1.0
 */

export interface TrendDataPoint {
  date: string;
  totalTransactions: number;
  flaggedCount: number;
  blockedCount: number;
  fraudRate: number;
}